requirements.txt
requests==2.28.1
retry==0.9.2
pyyaml==6.0
influxdb-client==1.33.0
hyundai_kia_connect_api==1.34.2
cachetools==5.2.0